Updating to HandBrake 4024svn. This should fix problems with corrupted mpeg2 preview images and encodes from mpeg2 sources reporting the wrong progress and reporting as failed. Fixed problem where target sizes over 2048 MB don't work. Updated version numbers.

+++ HandBrakeInterop/HandBrakeInterop/HandBrakeInstance.cs
@@ -290,7 +290,7 @@
 		/// <returns>The video bitrate in kbps.</returns>
 		public int CalculateBitrate(EncodeJob job, int sizeMB, double overallSelectedLengthSeconds = 0)
 		{
-			long availableBytes = sizeMB * 1024 * 1024;
+			long availableBytes = ((long) sizeMB) * 1024 * 1024;
 
 			EncodingProfile profile = job.EncodingProfile;
 			Title title = this.GetTitle(job.Title);
